requirements.txt
openai==0.28.1
python-dotenv>=0.19.0
streamlit>=1.22.0
python-docx>=0.8.11
PyPDF2>=3.0.0
